requirements.txt
Flask==3.0.2
Flask-Limiter==3.7.0
Flask-Talisman==1.1.0
PyJWT==2.8.0
requests==2.32.3
python-dotenv==1.0.1
redis==5.0.4 